refactoring + deleteToken row when email activated
