+++ src/main/java/docSharing/service/AuthService.java
@@ -5,32 +5,46 @@ import docSharing.entities.User;
 import docSharing.entities.VerificationToken;
 import docSharing.repository.UserRepository;
 import docSharing.repository.VerificationTokenRepository;
+import docSharing.utils.emailActivation.OnRegistrationCompleteEvent;
 import docSharing.utils.Utils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
 import java.sql.SQLDataException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class AuthService {
     @Autowired
     private final UserRepository userRepository;
-
     @Autowired
-    private VerificationTokenRepository tokenRepository;
+    private final VerificationTokenRepository tokenRepository;
+    @Autowired
+    ApplicationEventPublisher eventPublisher;
     static HashMap<User, String> mapUserTokens = new HashMap<>();
 
     private static final Logger logger = LogManager.getLogger(AuthService.class.getName());
 
-    public AuthService(UserRepository userRepository) {
+    public AuthService(UserRepository userRepository, VerificationTokenRepository tokenRepository) {
         this.userRepository = userRepository;
+        this.tokenRepository = tokenRepository;
     }
 
+
+
+//    public User createUser(UserRequest userRequest) throws SQLDataException {
+//        logger.info("in createUser");
+//
+//        if(userRepository.findByEmail(userRequest.getEmail())!=null){
+//            throw new SQLDataException(String.format("Email %s exists in users table", userRequest.getEmail()));
+//        }
+//
+//        logger.debug(userRequest);
+//        return userRepository.save(new User(userRequest.getName(), userRequest.getEmail(), userRequest.getPassword()));
+//    }
 
     public User createUser(UserRequest userRequest) throws SQLDataException {
         logger.info("in createUser");
@@ -40,7 +54,12 @@ public class AuthService {
         }
 
         logger.debug(userRequest);
-        return userRepository.save(new User(userRequest.getName(), userRequest.getEmail(), userRequest.getPassword()));
+        User userCreated = new User(userRequest.getName(), userRequest.getEmail(), userRequest.getPassword());
+        return userRepository.save(userCreated);
+
+//        String token = UUID.randomUUID().toString();
+//        createVerificationToken(userCreated, token);
+//        return userCreated;
     }
 
     public Optional<String> login(UserRequest userRequest) {
@@ -87,6 +106,7 @@ public class AuthService {
 
     public void createVerificationToken(User user, String token) {
         VerificationToken myToken = new VerificationToken(token, user);
+        logger.debug(myToken);
         tokenRepository.save(myToken);
     }
 
@@ -116,5 +136,13 @@ public class AuthService {
                 user.setEmail(newEmail);
             }
         }
+    }
+
+    public void publishRegistrationEvent(User createdUser, Locale locale, String appUrl  ) {
+        eventPublisher.publishEvent(new OnRegistrationCompleteEvent(createdUser, locale, appUrl));
+    }
+
+    public void deleteVerificationToken(String token) {
+        tokenRepository.deleteByToken(token);
     }
 }
