+++ src/main/java/docSharing/utils/emailActivation/RegistrationListener.java
@@ -1,7 +1,9 @@
-package docSharing.utils;
+package docSharing.utils.emailActivation;
 
 import docSharing.entities.User;
 import docSharing.service.AuthService;
+import docSharing.utils.emailActivation.GMailer;
+import docSharing.utils.emailActivation.OnRegistrationCompleteEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
