+++ src/main/resources/client_secret_695322809754-onh0sumimllsus0b2qotjfts1d9diflr.apps.googleusercontent.com.json
@@ -1,1 +1,13 @@
-{"installed":{"client_id":"695322809754-onh0sumimllsus0b2qotjfts1d9diflr.apps.googleusercontent.com","project_id":"share-doc-369209","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost"]}}
+{
+  "installed": {
+    "client_id": "695322809754-onh0sumimllsus0b2qotjfts1d9diflr.apps.googleusercontent.com",
+    "project_id": "share-doc-369209",
+    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+    "token_uri": "https://oauth2.googleapis.com/token",
+    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+    "client_secret": "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",
+    "redirect_uris": [
+      "http://localhost"
+    ]
+  }
+}
